+++ src/bsp/platform/stm32fx/xi_bsp_io_fs_stm32fx.c
@@ -10,7 +10,7 @@
 #include <stdio.h>
 
 xi_state_t xi_bsp_io_fs_open( const char* const resource_name,
-                              const uint32_t size,
+                              const size_t size,
                               const xi_fs_open_flags_t open_flags,
                               xi_fs_resource_handle_t* resource_handle_out )
 {
